# cabinet-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CabinetPro API",
        "description": "API REST de gestion de cabinet d'avocat — CabinetPro.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/dossiers": {
            "get": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Liste des dossiers",
                "operationId": "3639c07930436dd8ce5efde3abb52d4b",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type_droit",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priorite",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagination des dossiers"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Créer un dossier",
                "operationId": "464dbe49079bf88a59ebcb47a69596ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dossier"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dossier créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/dossiers/stats": {
            "get": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Statistiques des dossiers",
                "operationId": "dfc677316fa473b3e3c8825452253e89",
                "responses": {
                    "200": {
                        "description": "Statistiques"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/dossiers/{dossier}": {
            "get": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Détail d'un dossier",
                "operationId": "23ebf728f29d12d57f70a5ebe4df6018",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dossier"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Modifier un dossier",
                "operationId": "d0c471e016c5c88320fe22611b08029d",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dossier"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dossier modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Supprimer un dossier",
                "operationId": "c769a8a231e5aaff9f1b492acabe042c",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Dossier supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/dossiers/{dossier}/statut": {
            "patch": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Changer le statut d'un dossier",
                "operationId": "6a05909149bfce0cd754c438661dff36",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "statut"
                                ],
                                "properties": {
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "en_cours",
                                            "en_revision",
                                            "en_attente",
                                            "cloture"
                                        ]
                                    },
                                    "avancement": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dossier mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/dossiers/{dossier}/avocats": {
            "post": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Attacher ou détacher des avocats",
                "operationId": "c9aac39619e001102abd127bdca3d07e",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "avocat_ids"
                                ],
                                "properties": {
                                    "avocat_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avocats synchronisés"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/dossiers/{dossier}/activites": {
            "get": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Historique / audit d'un dossier",
                "operationId": "d1dc7ce8f144070295a162b30064dcaa",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Journal d'audit"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Dossier": {
                "required": [
                    "client_id",
                    "titre",
                    "type_droit",
                    "date_ouverture"
                ],
                "properties": {
                    "client_id": {
                        "type": "integer"
                    },
                    "titre": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "type_droit": {
                        "type": "string"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "en_cours",
                            "en_revision",
                            "en_attente",
                            "cloture"
                        ]
                    },
                    "priorite": {
                        "type": "string",
                        "enum": [
                            "normale",
                            "haute",
                            "urgente"
                        ]
                    },
                    "avancement": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 0
                    },
                    "montant": {
                        "description": "Montant en Ariary (entier)",
                        "type": "integer",
                        "nullable": true
                    },
                    "date_ouverture": {
                        "type": "string",
                        "format": "date"
                    },
                    "date_cloture": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "avocat_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}